requirements.txt
pandas
numpy
alpaca-trade-api
sqlalchemy
psycopg2-binary
pyyaml
lightgbm
scikit-learn
pyqlib
streamlit
matplotlib
plotly
python-dotenv
pydantic
pytest
schedule
requests
loguru

